feat(knex): add information whether workflow is latest
